remove roi head files
