+++ mmdet/models/tracking_heads/__init__.py
@@ -1,10 +1,5 @@
 # Copyright (c) OpenMMLab. All rights reserved.
 from .quasi_dense_embed_head import QuasiDenseEmbedHead
 from .quasi_dense_track_head import QuasiDenseTrackHead
-from .roi_embed_head import RoIEmbedHead
-from .roi_track_head import RoITrackHead
 
-__all__ = [
-    'RoITrackHead', 'RoIEmbedHead', 'QuasiDenseEmbedHead',
-    'QuasiDenseTrackHead'
-]
+__all__ = ['QuasiDenseEmbedHead', 'QuasiDenseTrackHead']
